requirements.txt
streamlit
earthengine-api
geemap
streamlit-folium
google-auth
